TextToImage-master/TextToImage-master/requirements.txt
Pillow==6.2.1

textwrap3==0.9.2
